Change a bunch of explicit value type conversions to implicit, and other changes to help reduce the chance of getting the wrong overload when using the property sync. Fixes propertysync for ActionDisplayID.

+++ netgore/trunk/NetGore/Stats/ModStatHelper.cs
@@ -18,11 +18,10 @@ namespace NetGore
         /// <param name="modders">The <see cref="IModStatContainer{TStatType}"/>s to use for calculating the modified
         /// stat value.</param>
         /// <returns>The modified stat value.</returns>
-        public static int Calculate<TStatType>(IStatCollection<TStatType> baseStats, TStatType statType,
-                                               params IModStatContainer<TStatType>[] modders)
+        public static int Calculate<TStatType>(IStatCollection<TStatType> baseStats, TStatType statType, params IModStatContainer<TStatType>[] modders)
             where TStatType : struct, IComparable, IConvertible, IFormattable
         {
-            int value = baseStats[statType];
+            int value = (int)baseStats[statType];
 
             if (modders == null || modders.Length == 0)
                 return value;
@@ -44,11 +43,10 @@ namespace NetGore
         /// <param name="modder1">The first <see cref="IModStatContainer{TStatType}"/>s to use for calculating
         /// the modified stat value.</param>
         /// <returns>The modified Stat value.</returns>
-        public static int Calculate<TStatType>(IStatCollection<TStatType> baseStats, TStatType statType,
-                                               IModStatContainer<TStatType> modder1)
+        public static int Calculate<TStatType>(IStatCollection<TStatType> baseStats, TStatType statType, IModStatContainer<TStatType> modder1)
             where TStatType : struct, IComparable, IConvertible, IFormattable
         {
-            int value = baseStats[statType];
+            int value = (int)baseStats[statType];
 
             value += modder1.GetStatModBonus(statType);
 
@@ -70,7 +68,7 @@ namespace NetGore
                                                IModStatContainer<TStatType> modder1, IModStatContainer<TStatType> modder2)
             where TStatType : struct, IComparable, IConvertible, IFormattable
         {
-            int value = baseStats[statType];
+            int value = (int)baseStats[statType];
 
             value += modder1.GetStatModBonus(statType);
             value += modder2.GetStatModBonus(statType);
@@ -92,11 +90,10 @@ namespace NetGore
         /// the modified stat value.</param>
         /// <returns>The modified stat value.</returns>
         public static int Calculate<TStatType>(IStatCollection<TStatType> baseStats, TStatType statType,
-                                               IModStatContainer<TStatType> modder1, IModStatContainer<TStatType> modder2,
-                                               IModStatContainer<TStatType> modder3)
+                                               IModStatContainer<TStatType> modder1, IModStatContainer<TStatType> modder2, IModStatContainer<TStatType> modder3)
             where TStatType : struct, IComparable, IConvertible, IFormattable
         {
-            int value = baseStats[statType];
+            int value = (int)baseStats[statType];
 
             value += modder1.GetStatModBonus(statType);
             value += modder2.GetStatModBonus(statType);
@@ -125,7 +122,7 @@ namespace NetGore
                                                IModStatContainer<TStatType> modder3, IModStatContainer<TStatType> modder4)
             where TStatType : struct, IComparable, IConvertible, IFormattable
         {
-            int value = baseStats[statType];
+            int value = (int)baseStats[statType];
 
             value += modder1.GetStatModBonus(statType);
             value += modder2.GetStatModBonus(statType);
